+++ CHANGES.md
@@ -11,10 +11,84 @@
 
 ## develop
 
+- [UPDATE] Github Actions の `macos-12` を `macos-14` に変更する
+  - @miosakuma
+
+## 2024.4.0 (2024-07-29)
+
+- [CHANGE] `--sora-dir`, `--sora-args` を `--local-sora-cpp-sdk-dir` と `--local-sora-cpp-sdk-args` に変更する
+  - @melpon
+- [CHANGE] `--webrtc-build-dir`, `--webrtc-build-args` を `--local-webrtc-build-dir` と `--local-webrtc-build-args` に変更する
+  - @melpon
+- [UPDATE] Sora C++ SDK を `2024.7.0` に上げる
+  - Sora C++ SDK に追従して H.264 専用となっていた `NvCodecH264Encoder` を `NvCodecVideoEncoder` に変更する
+  - WEBRTC_BUILD_VERSION を `m127.6533.1.1` にアップデート
+  - BOOST_VERSION を 1.85.0 にアップデート
+  - CMAKE_VERSION を 3.29.6 にアップデート
+  - @melpon @torikizi
+- [UPDATE] Github Actions で利用する Windows を 2022 にあげる
+  - @torikizi
+- [ADD] Ubuntu 22.04 でのビルドに対応する
+  - 単にローカルビルド可能にしただけで、リリースには含めていない
+  - @melpon
+- [ADD] WebSocket での接続時に User-Agent を設定する
+  - @melpon
+
+## 2024.3.0 (2024-04-18)
+
+- [UPDATE] Sora C++ SDK を `2024.6.1` に上げる
+  - Intel VPL H.265 対応（ Windows, Ubuntu 22.04 ）
+  - WEBRTC_BUILD_VERSION を `m122.6261.1.0` にあげる
+    - Ubuntu のビルドを通すために、 __assertion_handler というファイルをコピーする処理を追加した
+  - @miosakuma @torikizi
+- [UPDATE] run.py に定義されていた関数を buildbase.py に移動する
+  - @melpon
+- [UPDATE] Github Actions の actions/cache , actions/upload-artifact , actions/download-artifact をアップデート
+  - Node.js 16 の Deprecated に伴うアップデート
+    - actions/cache@v3 から actions/cache@v4 にアップデート
+    - actions/upload-artifact@v3 から actions/upload-artifact@v4 にアップデート
+    - actions/download-artifact@v3 から actions/download-artifact@v4 にアップデート
+  - @torikizi
+- [UPDATE] 対応プラットフォームに Ubuntu 22.04 x86_64 を追加する
+  - コード変更はなし、Ubuntu 20.04 版の Unity SDK で Ubuntu 22.04 も動作することを確認した
+  - @torikizi
+
+## 2024.2.0 (2024-03-13)
+
+- [CHANGE] Lyra を削除
+  - Sora.Config から AudioCodecLyraBitrate を削除
+  - Sora.Config から AudioCodecLyraUsedtx を削除
+  - Sora.Config から CheckLyraVersion を削除
+  - Sora.AudioCodecType から LYRA を削除
+  - VERSION ファイルから LYRA_VERSION を削除
+  - @melpon
+- [CHANGE] AudioOutputHelper を IAudioOutputHelper に変更
+  - Android のハンズフリー機能の追加に伴い、AudioOutputHelper を IAudioOutputHelper に変更
+  - これは破壊的な変更になるため、既存の AudioOutputHelper を利用している場合は、IAudioOutputHelper に変更が必要になる
+  - 以下のような理由から IAudioOutputHelper へ変更
+  - iOS と Android の異なるハンズフリー機能を同一の実装で呼び出せるように、現状のハンズフリー機能の呼び出し方法を変更
+    - Sora.AudioOutputHelper の生成を Factory にて行う
+    - AudioOutputHelper(Action onChangeRoute) から AudioOutputHelperFactory.create(Action onChangeRoute) に変更
+    - Sora.AudioOutputHelper は interface となるため Sora.IAudioOutputHelper に変更
+  - @tnoho
+- [UPDATE] Sora C++ SDK を `2024.4.0` に上げる
+  - Sora C++ SDK 2024.3.1 アップデート時に関連するライブラリもアップデート
+  - WEBRTC_BUILD_VERSION を `m121.6167.3.0` にアップデート
+  - BOOST_VERSION を 1.84.0 にアップデート
+  - CMAKE_VERSION を 3.28.1 にアップデート
+  - @melpon @miosakuma @torikizi
+- [ADD] Android 向けのハンズフリー機能を追加
+  - `Sora.cs` に Android 向けのハンズフリー機能を追加
+  - iOS と同様に Android でもハンズフリー機能の利用を可能にする
+  - @tnoho
+- [ADD] run.py に --sora-dir と --sora-args オプションを追加
+  - @melpon
+- [ADD] run.py に --webrtc-build-dir と --webrtc-build-args オプションを追加
+  - @melpon
+
 ## 2024.1.0 (2024-01-22)
 
 - [UPDATE] Sora C++ SDK を `2024.1.0` に上げる
-  - Sora C++ SDK 2024.1.0 アップデート時に関連するライブラリもアップデート
   - WEBRTC_BUILD_VERSION を `m120.6099.1.2` にアップデート
   - @torikizi
 - [ADD] VideoCodecType に H265 を追加